+++ src/main/y/information/translator/run-length.cpp
@@ -34,7 +34,7 @@ namespace upsylon {
             void Framework:: EmitDifferentTo(sequence<char> &S,
                                              IOBits         &Q,
                                              const uint8_t *cache,
-                                             const uint8_t  count,
+                                             const size_t   count,
                                              QEncoder      &characterEncoder,
                                              QEncoder      &repeatingEncoder,
                                              QEncoder      &differentEncoder)
